Add shared velocitas_lib
