+++ runtime-k3d/src/runtime/controlplane.py
@@ -12,11 +12,14 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+# flake8: noqa: E402 module level import
 import os
 import subprocess
 from typing import List
-
-from lib import get_script_path, require_env
+from pathlib import Path
+import sys
+sys.path.append(os.path.join(Path(__file__).parents[2], "velocitas_lib"))
+from velocitas_lib import require_env, get_script_path
 from yaspin.core import Yaspin
