+++ runtime-k3d/src/runtime/runtime.py
@@ -12,12 +12,16 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+# flake8: noqa: E402 module level import
 import subprocess
-
 from gen_helm import gen_helm
-from lib import get_services, parse_service_config
+from lib import parse_service_config
 from yaspin.core import Yaspin
-
+import os
+from pathlib import Path
+import sys
+sys.path.append(os.path.join(Path(__file__).parents[2], "velocitas_lib"))
+from velocitas_lib import get_cache_data, get_services
 
 def is_runtime_installed() -> bool:
     return (
@@ -49,6 +53,7 @@ def retag_docker_images():
 
 
 def create_vspec_config():
+    get_cache_data()["vspec_file_path"]
     subprocess.check_call([
         "kubectl",
         "create",
