+++ runtime-local/src/run-vehicle-app.py
@@ -12,12 +12,17 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
+# flake8: noqa: E402 module level import
 import argparse
 import subprocess
+import os
+from pathlib import Path
 import sys
+sys.path.append(os.path.join(Path(__file__).parents[2], "velocitas_lib"))
+from velocitas_lib import get_services
 from typing import Optional
 
-from lib import MiddlewareType, get_middleware_type, get_services, get_dapr_sidecar_args
+from lib import MiddlewareType, get_middleware_type, get_dapr_sidecar_args
 
 
 def get_dapr_app_id(service_id: str) -> str:
